chore: Rename GithubUrlDto to ReadmeItemDto
